refactor: move actions to lib/actions and update imports
